Initial port of KubeClient.Extensions.DataProtection implementation
This code has been ported from https://github.com/rriverak/KubeClient.Extensions.DataProtection (thanks, @rriverak!).
Note that, at this, stage, it is a work in progress (we need to decide on strategies for thread-safety and handling of common error conditions (such as the underlying Secret being deleted).

Relates to tintoy/dotnet-kube-client#94.